requirements/base.txt
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
pydantic==2.6.1
pyyaml==6.0.1
joblib==1.3.2
python-dotenv==1.0.0
requests==2.31.0